# perf(node/fs): speedup fs.readdir + async