--- a/ext/web/internal.d.ts
+++ b/ext/web/internal.d.ts
@@ -43,6 +43,7 @@ declare module "ext:deno_web/00_infra.js" {
   function forgivingBase64Decode(data: string): Uint8Array;
   function forgivingBase64UrlEncode(data: Uint8Array | string): string;
   function forgivingBase64UrlDecode(data: string): Uint8Array;
+  function pathFromURL(urlOrPath: string | URL): string;
   function serializeJSValueToJSONString(value: unknown): string;
 }