--- a/tests/unit_node/_fs/_fs_dirent_test.ts
+++ b/tests/unit_node/_fs/_fs_dirent_test.ts
@@ -1,0 +1,5 @@
+// Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
+
+// Empty file to make check_test_glob happy. The Dirent API
+// should not be consumed directly, but rather as a result of
+// fs.readdir and fs.readdirSync