Whilst working on Node compat issues I noticed that we could make the family of `readdir*` functions a lot faster

- Avoid shape transition when `withFileTypes` option is used
- Add fast path for usages without `withFileTypes` to skip allocations on both the JS and Rust side
- Call ops directly instead of layering on top of `Deno.*` JS APIs.

| Benchmark | main (iter/s) | PR (iter/s) |
|---|--:|--:|
| fs.readdir | 1,141.0 | 2,287.8 |
| fs.readdirSync | 1,434.9 | 2,406.8 |
| fs.readdir + withFileTypes | 1,101.0 | 1,407.2 |
| fs.readdirSync + withFileTypes | 1,372.8 | 1,481.9 |