@marvinhagemeister 2024-07-11
I've removed the old dirent tests because they were constructing `Dirent` instances directly which is disallowed in node types. Instead, I've ported those tests to two actual `readdir` ones that assert based on the actual result.